refactor: modify the handling of indexes on table "..."
In the run function, the treatment of indexes in the "..." table was changed because we will no longer ignore the first index in the other functions. The index always corresponded to the command that was going to be executed. Therefore, within the command function itself, it was irrelevant to have the command name as the first index.